requirements.txt
fastapi~=0.115.0
uvicorn
sqlalchemy~=2.0.35
psycopg2-binary
logging~=0.4.9.6
pydantic~=2.9.2
python-dotenv
transformers~=4.47.0
python-dotenv~=1.0.1
numpy~=2.0.2
requests~=2.32.3
gitpython